requirements.txt
asgiref==3.8.1
autopep8==2.3.2
bcrypt==4.2.0
cffi==1.17.1
coverage==7.9.1
cryptography==43.0.1
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
django-utils-six==2.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
iniconfig==2.1.0
packaging==25.0
paramiko==3.5.0
pillow==11.1.0
pluggy==1.6.0
pycodestyle==2.12.1
pycparser==2.22
Pygments==2.19.2
PyJWT==2.9.0
PyNaCl==1.5.0
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
six==1.17.0
sqlparse==0.5.3
